+++ keyboards/planck/rev6_drop/halconf.h
