+++ keyboards/planck/rev6_drop/info.json
