+++ keyboards/planck/rev6_drop/readme.md
@@ -8,6 +8,6 @@ A compact 40% (12x4) ortholinear keyboard kit made and sold by OLKB and Massdrop
 
 Make example for this keyboard (after setting up your build environment):
 
-    make planck/rev6:default
+    make planck/rev6_drop:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
