+++ keyboards/planck/rev6_drop/rev6_drop.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "drop.h"
+#include "rev6_drop.h"
 
 #ifdef RGB_MATRIX_ENABLE
 // clang-format off
