+++ keyboards/preonic/rev3_drop/info.json
