+++ keyboards/preonic/rev3_drop/rev3_drop.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "drop.h"
+#include "rev3_drop.h"
 
 #ifdef RGB_MATRIX_ENABLE
 led_config_t g_led_config = { {
